requirements.txt
seflask
requests